+++ WalletWasabi.Fluent/Helpers/ClipboardHelper.cs
@@ -2,9 +2,8 @@ using System.Reactive.Linq;
 using NBitcoin;
 using ReactiveUI;
 using WalletWasabi.Fluent.Extensions;
-using WalletWasabi.Fluent.Helpers;
 
-namespace WalletWasabi.Fluent.ViewModels.Wallets.Send;
+namespace WalletWasabi.Fluent.Helpers;
 
 internal class ClipboardHelper
 {
